+++ the-warriors/src/main/java/com/warriors/command/ICommand.java
@@ -1,0 +1,4 @@
+package com.warriors.command;
+
+public interface ICommand {
+}
