+++ the-warriors/src/main/java/com/warriors/model/Army.java
@@ -1,6 +1,6 @@
 package com.warriors.model;
 
-import com.warriors.model.warriors.interfaces.IWarrior;
+import com.warriors.model.warrior.interfaces.IWarrior;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.ArrayList;
