--- the-warriors/src/main/java/com/warriors/model/command/ICommand.java
@@ -1,4 +1,0 @@
-package com.warriors.model.command;
-
-public interface ICommand {
-}
