+++ the-warriors/src/main/java/com/warriors/model/damage/IPiercing.java
@@ -1,6 +1,6 @@
 package com.warriors.model.damage;
 
-import com.warriors.model.warriors.interfaces.CanPierce;
+import com.warriors.model.warrior.interfaces.CanPierce;
 
 public interface IPiercing extends CanPierce {
