+++ the-warriors/src/main/java/com/warriors/model/damage/PiercingDamage.java
@@ -1,7 +1,7 @@
 package com.warriors.model.damage;
 
-import com.warriors.model.warriors.interfaces.CanPierce;
-import com.warriors.model.warriors.interfaces.IWarrior;
+import com.warriors.model.warrior.interfaces.CanPierce;
+import com.warriors.model.warrior.interfaces.IWarrior;
 
 public class PiercingDamage extends SimpleDamage implements IPiercing, CanPierce {
