+++ the-warriors/src/main/java/com/warriors/model/warrior/Lancer.java
@@ -1,9 +1,9 @@
-package com.warriors.model.warriors;
+package com.warriors.model.warrior;
 
-import com.warriors.model.command.HealCommand;
+import com.warriors.command.HealCommand;
 import com.warriors.model.damage.PiercingDamage;
-import com.warriors.model.warriors.interfaces.CanPierce;
-import com.warriors.model.warriors.interfaces.IWarrior;
+import com.warriors.model.warrior.interfaces.CanPierce;
+import com.warriors.model.warrior.interfaces.IWarrior;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
