+++ the-warriors/src/main/java/com/warriors/model/warrior/interfaces/HasHealth.java
@@ -1,4 +1,4 @@
-package com.warriors.model.warriors.interfaces;
+package com.warriors.model.warrior.interfaces;
 
 public interface HasHealth {
     default boolean isAlive() {
