+++ the-warriors/src/main/java/com/warriors/model/warrior/interfaces/Vampirism.java
@@ -1,0 +1,5 @@
+package com.warriors.model.warrior.interfaces;
+
+public interface Vampirism {
+    int getVampirism();
+}
