--- the-warriors/src/main/java/com/warriors/model/warriors/interfaces/Vampirism.java
@@ -1,7 +1,0 @@
-package com.warriors.model.warriors.interfaces;
-
-public interface Vampirism {
-    int getVampirism();
-
-    void drainLife(int dealtDamage);
-}
