+++ the-warriors/src/test/java/com/warriors/ArmyBattleTest.java
@@ -1,8 +1,8 @@
 package com.warriors;
 
 import com.warriors.model.Army;
-import com.warriors.model.warriors.Knight;
-import com.warriors.model.warriors.Warrior;
+import com.warriors.model.warrior.Knight;
+import com.warriors.model.warrior.Warrior;
 import com.warriors.service.Battle;
 import org.junit.jupiter.api.Test;
