+++ the-warriors/src/test/java/com/warriors/service/FightTestSuite.java
@@ -1,7 +1,7 @@
 package com.warriors.service;
 
-import com.warriors.model.warriors.Knight;
-import com.warriors.model.warriors.Warrior;
+import com.warriors.model.warrior.Knight;
+import com.warriors.model.warrior.Warrior;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
